OpenLabelBackend/OpenLabelDB/requirements.txt
pymongo==4.3.3
python-dotenv==1.0.0
bcrypt==4.0.1
pyfiglet==0.8.post1
tabulate==0.9.0